requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.6
django-crispy-forms==2.3
djangorestframework==3.15.2
parameterized==0.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
